pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oculus-vorago"
version = "0.1.0"
description = "OculusVorago — a highly generic, scalable ETL pipeline for the Oculus intelligence platform."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["etl", "kafka", "pipeline", "oculus", "intelligence"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "confluent-kafka>=2.4; sys_platform != 'emscripten'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
vorago = "main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["vorago*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = true
log_cli_level = "DEBUG"
